features/document_analysis/requirements.txt
# Document Analysis Feature Requirements

# File Upload and Processing
aiofiles>=23.2.1
python-multipart>=0.0.6

# Document Parsing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0
striprtf>=0.0.26
chardet>=5.2.0

# Task Queue
redis>=5.0.1
celery>=5.3.4
kombu>=5.3.4

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1

# API and WebSocket
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0

# Utilities
pydantic>=2.5.0
python-magic>=0.4.27
pillow>=10.1.0